added appSettings to appStarter.R to give the user more control without editing the surce code files

+++ packageManager.R
@@ -1,9 +1,15 @@
-apapp.pm.requiredPackages <- c("shiny","shinythemes","shinyjs","shinycssloaders","ggplot2","Rcpp","highcharter","xts")
+apapp.pm.getRequiredPackages <- function(useHighcharts, useCpp)  {
+  requiredPackages <- c("shiny","shinythemes","shinyjs","shinycssloaders","ggplot2")
+  if(useCpp) requiredPackages <- c(requiredPackages,"Rcpp")
+  if(useHighcharts) requiredPackages <- c(requiredPackages,"highcharter","xts")
+  return(requiredPackages)
+}
 
-apapp.pm.installMissingPackages <- function(rPackageRepository = "https://cran.rstudio.com/"){
+apapp.pm.installMissingPackages <- function(rPackageRepository = "https://cran.rstudio.com/",
+                                            requiredPackages){
   availablePackages <- unname(utils::installed.packages()[,1])
   
-  needToInstall <- apapp.pm.requiredPackages[!(apapp.pm.requiredPackages %in% availablePackages)]
+  needToInstall <- requiredPackages[!(requiredPackages %in% availablePackages)]
   
   if(!length(needToInstall)) {
     print("Alle nötigen Packages sind bereits installiert.")
@@ -18,7 +24,7 @@ apapp.pm.installMissingPackages <- function(rPackageRepository = "https://cran.r
   print("Alle nötigen Packages wurden installiert.")
 }
 
-apapp.pm.loadRequiredPackages <- function(requiredPackages = apapp.pm.requiredPackages) {
+apapp.pm.loadRequiredPackages <- function(requiredPackages) {
   for(p in requiredPackages) library(p, character.only = TRUE)
 }
 
@@ -50,7 +56,7 @@ apapp.pm.sourceNeededFiles <- function(gitDirectoryPath, sourceCppToo = TRUE) {
 }
 
 
-apapp.pm.sourceNeededFilesFromRepo <- function() {
+apapp.pm.sourceNeededFilesFromRepo <- function(useCpp) {
   rFiles <- c("https://raw.githubusercontent.com/JonasKir97/aparch_app/master/acogarchSimulation.R",
               "https://raw.githubusercontent.com/JonasKir97/aparch_app/master/acogarchSimulationAppHelpers.R",
               "https://raw.githubusercontent.com/JonasKir97/aparch_app/master/guiHelpers.R",
@@ -62,18 +68,21 @@ apapp.pm.sourceNeededFilesFromRepo <- function() {
     base::source(file = rFile, encoding = "UTF-8")
   }
   
-  cppFiles <- c("https://raw.githubusercontent.com/JonasKir97/aparch_app/master/simulateDiscreteAPARCH11inCPP.cpp")
-  
-  sourceCppFromGithub <- function(url) {
-    tf <- tempfile(fileext = ".cpp") 
-    utils::download.file(url, tf, quiet=TRUE) 
-    Rcpp::sourceCpp(file = tf)
-    file.remove(tf)
+  if(useCpp) {
+    cppFiles <- c("https://raw.githubusercontent.com/JonasKir97/aparch_app/master/simulateDiscreteAPARCH11inCPP.cpp")
+    
+    sourceCppFromGithub <- function(url) {
+      tf <- tempfile(fileext = ".cpp") 
+      utils::download.file(url, tf, quiet=TRUE) 
+      Rcpp::sourceCpp(file = tf)
+      file.remove(tf)
+    }
+    
+    for(cFile in cppFiles) {
+      print(paste0("Source cpp-Datei ",cFile,"."))
+      sourceCppFromGithub(cFile)
+    }
   }
-  
-  for(cFile in cppFiles) {
-    print(paste0("Source cpp-Datei ",cFile,"."))
-    sourceCppFromGithub(cFile)
-  }
+
   print("Sourcing done.")
 }
